Update README.md
Normally, an FTP client app can only upload files under one folder on the FTP site at a time. This is extremely cumbersome! This utility is used to upload an entire nested hierarchical local directory to an FTP site in one go.

The function is built by using Python `ftppretty` and `pathlib` Libraries.